Initial version

+++ leader-election/src/main/java/io/vepo/distributedsystem/leaderelection/LeaderElection.java
@@ -13,8 +13,12 @@ public class LeaderElection {
     public static void main(String[] args) throws Exception {
         logger.info("Starting leader election...");
         var rand = new Random();
-        var config = ClusterConfig.builder().id(UUID.randomUUID().toString()).multicastIp("224.0.0.1")
-                .multicastPort(4445).timeout(Duration.ofSeconds(5).plus(Duration.ofMillis(rand.nextInt(500) + 250)))
+        var config = ClusterConfig.builder()
+                .id(UUID.randomUUID().toString())
+                .multicastIp("224.0.0.1")
+                .multicastPort(4445)
+                .timeout(Duration.ofSeconds(5)
+                        .plus(Duration.ofMillis((long) rand.nextInt(500) + 250)))
                 .build();
         try (var controller = new ClusterController(config)) {
             controller.joinCluster();
